Tests/LibWeb/Text/expected/css-has-invalidation/has-feature-metadata-buckets.txt

[add unrelated class]
  hasInvalidationMetadataCandidates: 0
  hasAncestorWalkInvocations: 0
  hasAncestorWalkVisits: 0
  styleInvalidations: 0
[add class referenced by one :has()]
  hasInvalidationMetadataCandidates: 1
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 0
  styleInvalidations: 0
[add class referenced by two :has() rules]
  hasInvalidationMetadataCandidates: 2
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 0
  styleInvalidations: 0
